+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-v2.3 (??? 2024)
+v2.3 (Oct 2024)
     * Moved plotting routines to standalone package ``mcplot``. Add
       DeprecationWarning to plotting routines.
     * Add DeprecationWarning to routines for screening with Elementary